completing import service: re-init configurations when after import

+++ src/main/java/com/idemia/tec/jkt/cardiotest/service/ExportImportService.java
@@ -16,6 +16,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
 
 @Service
@@ -23,11 +24,55 @@ public class ExportImportService {
 
     static Logger logger = Logger.getLogger(ExportImportService.class);
 
-    private String scriptsDirectory;
-    private File exportRunSettingsFile;
+    private String impScriptsDirectory;
+    private String expScriptsDirectory;
 
-    public void importSettings() {
-        // TODO
+    public boolean importSettings(File importFile, File projectDir, File advSaveVar) throws IOException {
+        // inflate zip file to tmp
+        File tmpDir = new File("tmp\\");
+        if (!tmpDir.exists()) if (tmpDir.mkdir()) logger.info("Created tmp directory in app root");
+        if (!unzip(importFile, tmpDir)) return false;
+
+        // create structure in project folder
+        impScriptsDirectory = projectDir.getAbsolutePath() + "\\scripts\\";
+        File scriptDir = new File(impScriptsDirectory);
+        if (!scriptDir.exists()) if (scriptDir.mkdir()) logger.info("Created directory " + scriptDir.getAbsolutePath());
+
+        // parse exported RunSettings (with ObjectMapper), and set as input from user: projectPath & advSaveVariablesPath
+        File exportRunSettingsFile = new File("tmp\\ex-run-settings.json");
+        ObjectMapper ersMapper = new ObjectMapper();
+        RunSettings rs = ersMapper.readValue(exportRunSettingsFile, RunSettings.class);
+        rs.setProjectPath(projectDir.getAbsolutePath());
+        rs.setAdvSaveVariablesPath(advSaveVar.getAbsolutePath());
+
+        // TODO get readerNumber from current and set to exported RunSettings*
+
+        // copy custom scripts (if any)
+        if (rs.getCustomScriptsSection1().size() > 0) copyScriptFromTmp(rs.getCustomScriptsSection1());
+        if (rs.getCustomScriptsSection2().size() > 0) copyScriptFromTmp(rs.getCustomScriptsSection2());
+        if (rs.getCustomScriptsSection3().size() > 0) copyScriptFromTmp(rs.getCustomScriptsSection3());
+
+        // save exported RunSettings to run-settings.json
+        File runSettingsFile = new File("run-settings.json");
+        ObjectMapper rsMapper = new ObjectMapper();
+        rsMapper.writerWithDefaultPrettyPrinter().writeValue(runSettingsFile, rs);
+
+        // clean tmp folder
+        List<File> tmpFiles = new ArrayList<>();
+        try (Stream<Path> walk = Files.walk(Paths.get("tmp"))) {
+            List<String> srcFileNames = walk.filter(Files::isRegularFile).map(Path::toString).collect(Collectors.toList());
+            for (String src : srcFileNames) tmpFiles.add(new File(src));
+        }
+        for (File tmpFile : tmpFiles) Files.deleteIfExists(tmpFile.toPath()); // delete all files in temporary folder
+        return true;
+    }
+
+    private void copyScriptFromTmp(List<CustomScript> customScripts) throws IOException {
+        for (CustomScript cs : customScripts) {
+            File source = new File("tmp\\" + cs.getCustomScriptName());
+            File target = new File(impScriptsDirectory + cs.getCustomScriptName());
+            Files.copy(source.toPath(), target.toPath(), StandardCopyOption.REPLACE_EXISTING);
+        }
     }
 
     public String exportSettings(RunSettings rs, File exportFile) throws IOException {
@@ -43,7 +88,7 @@ public class ExportImportService {
         if (!tmpDir.exists()) if (tmpDir.mkdir()) logger.info("Created tmp directory in app root");
 
         // settings file
-        exportRunSettingsFile = new File("tmp\\ex-run-settings.json");
+        File exportRunSettingsFile = new File("tmp\\ex-run-settings.json");
         ObjectMapper ersMapper = new ObjectMapper();
         try { ersMapper.writerWithDefaultPrettyPrinter().writeValue(exportRunSettingsFile, ers); }
         catch (IOException e) { logger.error("Failed copying settings: " + e.getMessage()); }
@@ -55,7 +100,7 @@ public class ExportImportService {
         catch (IOException e) { logger.error("Failed copying variables: " + e.getMessage()); }
 
         // custom scripts
-        scriptsDirectory = rs.getProjectPath() + "\\scripts\\";
+        expScriptsDirectory = rs.getProjectPath() + "\\scripts\\";
         if (rs.getCustomScriptsSection1().size() > 0) copyScriptToTmp(rs.getCustomScriptsSection1());
         if (rs.getCustomScriptsSection2().size() > 0) copyScriptToTmp(rs.getCustomScriptsSection2());
         if (rs.getCustomScriptsSection3().size() > 0) copyScriptToTmp(rs.getCustomScriptsSection3());
@@ -66,17 +111,15 @@ public class ExportImportService {
             List<String> srcFileNames = walk.filter(Files::isRegularFile).map(Path::toString).collect(Collectors.toList());
             for (String src : srcFileNames) srcFiles.add(new File(src));
         }
-        boolean zipOk = zip(srcFiles, exportFile);
+        if (!zip(srcFiles, exportFile)) return "Failed exporting settings.";
 
         for (File srcFile : srcFiles) Files.deleteIfExists(srcFile.toPath()); // delete all files in temporary folder
-
-        if (zipOk) return "Exported settings to " + exportFile.getAbsolutePath();
-        else return "Failed exporting settings.";
+        return "Exported settings to " + exportFile.getAbsolutePath();
     }
 
     private void copyScriptToTmp(List<CustomScript> customScripts) throws IOException {
         for (CustomScript cs : customScripts) {
-            File source = new File(scriptsDirectory + cs.getCustomScriptName());
+            File source = new File(expScriptsDirectory + cs.getCustomScriptName());
             File target = new File("tmp\\" + cs.getCustomScriptName());
             Files.copy(source.toPath(), target.toPath(), StandardCopyOption.REPLACE_EXISTING);
         }
@@ -104,6 +147,35 @@ public class ExportImportService {
             logger.error("Error zipping files: " + e.getMessage());
             return false;
         }
+    }
+
+    private boolean unzip(File zipFile, File destDir) {
+        byte[] buffer = new byte[1024];
+        try {
+            ZipInputStream zis = new ZipInputStream(new FileInputStream(zipFile.getAbsolutePath()));
+            ZipEntry zipEntry = zis.getNextEntry();
+            while (zipEntry != null) {
+                File newFile = newFile(destDir, zipEntry);
+                FileOutputStream fos = new FileOutputStream(newFile);
+                int len;
+                while ((len = zis.read(buffer)) > 0) fos.write(buffer, 0, len);
+                fos.close();
+                zipEntry = zis.getNextEntry();
+            }
+            return true;
+        } catch (IOException e) {
+            logger.error("Error unzipping file: " + e.getMessage());
+            return false;
+        }
+    }
+
+    private File newFile(File destinationDir, ZipEntry zipEntry) throws IOException {
+        File destFile = new File(destinationDir, zipEntry.getName());
+        String destDirPath = destinationDir.getCanonicalPath();
+        String destFilePath = destFile.getCanonicalPath();
+        if (!destFilePath.startsWith(destDirPath + File.separator))
+            throw new IOException("Entry is outside of the target dir: " + zipEntry.getName());
+        return destFile;
     }
 
 }
